+++ slashing-protection/src/main/java/tech/pegasys/web3signer/slashingprotection/dao/Validator.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.slashingprotection;
+package tech.pegasys.web3signer.slashingprotection.dao;
 
 import org.apache.tuweni.bytes.Bytes;
