+++ slashing-protection/src/test/java/tech/pegasys/web3signer/slashingprotection/dao/ValidatorsDaoTest.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.slashingprotection;
+package tech.pegasys.web3signer.slashingprotection.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -27,6 +27,10 @@ import org.jdbi.v3.testing.Migration;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+import tech.pegasys.web3signer.slashingprotection.BytesArgumentFactory;
+import tech.pegasys.web3signer.slashingprotection.BytesColumnMapper;
+import tech.pegasys.web3signer.slashingprotection.dao.Validator;
+import tech.pegasys.web3signer.slashingprotection.dao.ValidatorsDao;
 
 // This must be a junit4 for the JdbiRule to work
 public class ValidatorsDaoTest {
